EF repositories

+++ HotelProject.Repository/MicrosoftDataSQLClient/RoomRepository.cs
@@ -35,7 +35,7 @@ namespace HotelProject.Repository.MicrosoftDataSQLClient
                                 Name = !reader.IsDBNull(1) ? reader.GetString(1) : string.Empty,
                                 IsFree = !reader.IsDBNull(2) ? reader.GetBoolean(2) : false,
                                 HotelId = !reader.IsDBNull(3) ? reader.GetInt32(3) : 0,
-                                DailyPrice = !reader.IsDBNull(4) ? reader.GetDecimal(4) : 0
+                                DailyPrice = !reader.IsDBNull(4) ? reader.GetDouble(4) : 0
                             });
                         }
                     }
@@ -80,7 +80,7 @@ namespace HotelProject.Repository.MicrosoftDataSQLClient
                             result.Name = !reader.IsDBNull(1) ? reader.GetString(1) : string.Empty;
                             result.IsFree = !reader.IsDBNull(2) ? reader.GetBoolean(2) : false;
                             result.HotelId = !reader.IsDBNull(3) ? reader.GetInt32(3) : 0;
-                            result.DailyPrice = !reader.IsDBNull(4) ? reader.GetDecimal(4) : 0;
+                            result.DailyPrice = !reader.IsDBNull(4) ? reader.GetDouble(4) : 0;
                         }
                     }
                 }
